==> NOTEBOOKS/1.-preprocesing.ipynb <==
# Librerias

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder,MinMaxScaler, OrdinalEncoder
import random
pd.set_option('display.max_columns',None)

import warnings
warnings.filterwarnings("ignore")
from sklearn.preprocessing import OrdinalEncoder

import warnings
warnings.filterwarnings("ignore")

# fechas
from datetime import datetime

# gráficos
import seaborn as sns
import matplotlib.pyplot as plt

# preprocessing
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, Normalizer, MinMaxScaler
from sklearn.preprocessing import PolynomialFeatures

# algoritmos de clasificación
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression, RidgeClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from xgboost import XGBClassifier
from catboost import CatBoostClassifier

# algoritmos de regresión
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet, SGDRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from xgboost import XGBRegressor
from catboost import CatBoostRegressor

# métricas
import sklearn.metrics as metrics

# hiperparametrización
from sklearn.model_selection import GridSearchCV

# TensorFlow y tf.keras
import tensorflow as tf
from tensorflow import keras

import xgboost as xgb

from numpy import random
random.seed(42)

from sklearn.model_selection import train_test_split

# Configurar pandas para mostrar más valores
np.set_printoptions(threshold=np.inf)

# Cargar datos

# Usuario Windows carga datos desde aquí

""" file_path = "C:/Users/Herre/Desktop/commercial_activity_df.csv"
file_path_2 = "C:/Users/Herre/Desktop/products_df.csv"
file_path_3 = "C:/Users\Herre\Desktop\sociodemographic_df.csv"
"""

# Usuario Mac
# Ojo con duplicar las comillas

file_path = "/Users/ivan/Library/Mobile Documents/com~apple~CloudDocs/Software Engineer/TFM/Local_Nuclio_EasyMoney/data/commercial_activity_df.csv"
file_path_2 = "/Users/ivan/Library/Mobile Documents/com~apple~CloudDocs/Software Engineer/TFM/Local_Nuclio_EasyMoney/data/products_df.csv"
file_path_3 = "/Users/ivan/Library/Mobile Documents/com~apple~CloudDocs/Software Engineer/TFM/Local_Nuclio_EasyMoney/data/sociodemographic_df.csv"

df_comercial_activity = pd.read_csv(file_path)
df_products = pd.read_csv(file_path_2)
df_socialdemographic = pd.read_csv(file_path_3)

df_products.info()

# EDA: Análisis Exploratorio

## Quedarnos con las fechas más recientes

VEMOS QUE HAY CLIENTES QUE ESTÁN REPETIDOS, NOS QUEDAMOS CON SU INFORMACIÓN MAS RCIENTE, ES DECIR, QUE NOS QUEDAMOS CON UN SOLA "FILA" POR CLIENTE  (LA QUE TENGA LA FECHA MÁS RECIENTE, ESO EN TODOS LOS DATAFRAME Y LUEGO INTENTAMOS HACER UN MERGE, PRIMERO CON "PK_CID" Y LUEGO JUNTAMOS LOS 2 DF QUE QUEDAN A APRTIR DE LA COLUMNA "PK_PARTITION" PORQUE SE SUPONE QUE TENEMOS LAS MISMAS FECHAS EN AMBOS DF)

Hay que tener en cuenta que para unir df_comercial_activity y df_socialdemographic solo se puede hacer a través de la columna "pk_cid", por lo tanto, lo primero que hay que hacer es quedarnos con la info de cada cliente pero sin que esté repetida, nos quedamos con la información con fecha más reciente.

# Distribución por producto

Nos damos cuenta de que al hacer el modelo de que las variables 0 pertenecen al 99% de los datos, verificamos en el dataset original si es así. Efectivamente lo es.

# Revisar la distribución de las variables producto
df_products['emc_account'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

## EDA df_comercial_activity

df_comercial_activity.info()

df_comercial_activity.head()

columnas_categoricas = df_comercial_activity.select_dtypes(include=['object']).columns.tolist()
columnas_numericas = df_comercial_activity.select_dtypes(include=['int', 'float']).columns.tolist()

for col in columnas_categoricas:
    unique_values = df_comercial_activity[col].unique()
    print(f"Unique values in {col}: {unique_values}")

Convertir columnas tipo object a tipo datetime

df_comercial_activity['pk_partition'] = pd.to_datetime(df_comercial_activity['pk_partition'], errors='coerce', format='%Y-%m-%d')
df_comercial_activity['entry_date'] = pd.to_datetime(df_comercial_activity['entry_date'], errors='coerce', format='%Y-%m-%d')

### df_comercial_activity usuarios con fecha más reciente

df_comercial_activity["pk_cid"].value_counts()

# Quedarse con la transacción más reciente por usuario (pk_cid)
df_comercial_activity = df_comercial_activity.sort_values('pk_partition').groupby('pk_cid').tail(1).reset_index(drop=True)

df_comercial_activity["pk_cid"].value_counts()

df_comparison = df_comercial_activity[['pk_cid', 'pk_partition']]
df_comparison

## EDA df_products

df_products.info()

df_products.head()

# Configurar pandas para mostrar más valores
np.set_printoptions(threshold=np.inf)

df_products['pk_cid'].value_counts(normalize=True) * 100

columnas_categoricas_1 = df_products.select_dtypes(include=['object']).columns.tolist()
columnas_numericas_1 = df_products.select_dtypes(include=['int', 'float']).columns.tolist()

for col in columnas_categoricas_1:
    unique_values = df_products[col].unique()
    print(f"Unique values in {col}: {unique_values}")

Convertir columnas tipo object a tipo datetime

# Convertir la columna 'pk_partition' a tipo datetime
df_products['pk_partition'] = pd.to_datetime(df_products['pk_partition'], errors='coerce', format='%Y-%m-%d')

df_products['pk_partition'].unique()
#Vemos que se han hecho los cambios

### df_products usuarios con fecha más reciente

df_products["pk_cid"].value_counts()

# Quedarse con la transacción más reciente por usuario (pk_cid)
df_products = df_products.sort_values('pk_partition').groupby('pk_cid').tail(1).reset_index(drop=True)

df_products["pk_cid"].value_counts()

df_comparison_2 = df_products[['pk_cid', 'pk_partition']]
df_comparison_2

## EDA df_socialdemographic

df_socialdemographic.info()

df_socialdemographic.head()

columnas_categoricas_2 = df_socialdemographic.select_dtypes(include=['object']).columns.tolist()
columnas_numericas_2 = df_socialdemographic.select_dtypes(include=['int', 'float']).columns.tolist()

for col in columnas_categoricas_2:
    unique_values = df_socialdemographic[col].unique()
    print(f"Unique values in {col}: {unique_values}")

Convertir columnas tipo object en tipo datatime

# Convertir la columna 'pk_partition' a tipo datetime
df_socialdemographic['pk_partition'] = pd.to_datetime(df_socialdemographic['pk_partition'])

df_socialdemographic['pk_partition'].unique()
#Vemos que se han hecho los cambios

### df_socialdemographic usuarios con fecha más reciente

df_socialdemographic["pk_cid"].value_counts()

# Quedarse con la transacción más reciente por usuario (pk_cid)
df_socialdemographic = df_socialdemographic.sort_values('pk_partition').groupby('pk_cid').tail(1).reset_index(drop=True)

df_socialdemographic["pk_cid"].value_counts()

df_comparison_1 = df_socialdemographic[['pk_cid', 'pk_partition']]
df_comparison_1

## Análisis Univariable

### Análisis Univariable df_comercial_activity

#Primero vamos a hacer los histogramas de las variables numéricas.
import matplotlib.pyplot as plt

# Seleccionar solo las columnas numéricas
numeric_columns = df_comercial_activity.select_dtypes(include=[int, float]).columns

# Crear gráficos individuales
for col in numeric_columns:
    plt.figure(figsize=(10, 5))
    df_comercial_activity[col].hist(bins=50)
    plt.title(col)
    plt.show()

#Ahora crearemos histogramas para todas las columnas categóricas
categorical_columns = df_comercial_activity.select_dtypes(include=[object, 'bool'])
for column in categorical_columns:
    df_comercial_activity[column].value_counts().plot(kind='bar', title=column)
    plt.show()

### Análisis Univariable df_products

#Primero vamos a hacer los histogramas de las variables numéricas.
import matplotlib.pyplot as plt

# Seleccionar solo las columnas numéricas
numeric_columns_1 = df_products.select_dtypes(include=[int, float]).columns

# Crear gráficos individuales
for col in numeric_columns_1:
    plt.figure(figsize=(10, 5))
    df_products[col].hist(bins=50)
    plt.title(col)
    plt.show()

#Ahora crearemos histogramas para todas las columnas categóricas
categorical_columns_1 = df_products.select_dtypes(include=[object])
for column in categorical_columns_1:
    df_products[column].value_counts().plot(kind='bar', title=column)
    plt.show()

### Análisis Univariable df_socialdemographic

#Primero vamos a hacer los histogramas de las variables numéricas.
import matplotlib.pyplot as plt

# Seleccionar solo las columnas numéricas
numeric_columns_2 = df_socialdemographic.select_dtypes(include=[int, float]).columns

# Crear gráficos individuales
for col in numeric_columns_2:
    plt.figure(figsize=(10, 5))
    df_socialdemographic[col].hist(bins=50)
    plt.title(col)
    plt.show()

#Ahora crearemos histogramas para todas las columnas categóricas
categorical_columns_2 = df_socialdemographic.select_dtypes(include=[object, 'bool'])
for column in categorical_columns_2:
    df_socialdemographic[column].value_counts().plot(kind='bar', title=column)
    plt.show()

# Data Cleaning

#Ver duplicados de filas exactas

df_comercial_activity[df_comercial_activity.duplicated(keep=False)]

df_products[df_products.duplicated(keep=False)]

df_socialdemographic[df_socialdemographic.duplicated(keep=False)]

## NULOS

### Nulos df_comercial_activity

df_comercial_activity.info()

#Vemos los nulos que hay en cada columna
df_comercial_activity.isnull().sum()

# Como previamente hemos hecho una visualización de las columnas y sus variables ya tenemos una idea clara de las acciones que vamos a realizar con cada columna.

# Eliminamos la columna Unnamed

df_comercial_activity.drop(columns=['Unnamed: 0'], inplace=True)

df_comercial_activity = df_comercial_activity.dropna(subset=['entry_channel', 'entry_date', 'segment'])

#Una vez imputados los nulos agrupamos todas quellas variables que representan muy poco porcentaje sobre el total, de esta forma reducimos variables.
#Vamos a agrupar todas aquellas variables que representen el 7% o menos
# Calcular el porcentaje de cada categoría en entry_channel
porcentajes = df_comercial_activity['entry_channel'].value_counts(normalize=True)

# Identificar las categorías con porcentaje menor o igual al 7%
categorias_bajas = porcentajes[porcentajes <= 0.07].index

# Reemplazar las categorías de bajo porcentaje con "Otros"
df_comercial_activity['entry_channel'] = df_comercial_activity['entry_channel'].replace(categorias_bajas, 'Otros')

# Verificar el resultado
print(df_comercial_activity['entry_channel'].value_counts(normalize=True))

df_comercial_activity.reset_index(drop=True,inplace=True)

### Nulos df_products

df_products.isnull().sum()

#Eliminamos la columna Unnamed

df_products.drop(columns=['Unnamed: 0'], inplace=True)

df_products.reset_index(drop=True,inplace=True)

### Nulos df_socialdemographic

df_socialdemographic.isnull().sum()

#Eliminamos la columna Unnamed

df_socialdemographic.drop(columns=['Unnamed: 0'], inplace=True)

#Primero vamos a agrupar todas las variables menos (ES) de la columna "country_id" en una sola variable llamada Otros
# Agrupar todas las variables menos 'ES' en 'Otros' para la columna 'country_id'
df_socialdemographic['country_id'] = df_socialdemographic['country_id'].apply(lambda x: x if x == 'ES' else 'Otros')

# Verificar el resultado
df_socialdemographic['country_id'].value_counts()


df_socialdemographic['region_code'].value_counts()

#En esta base de datos hay columnas con pocos nulos así que los vamos a eliminar
df_socialdemographic = df_socialdemographic.dropna(subset=['region_code', 'gender'])

Nos hemos dado cuenta de que el nº de nulos en "region_code" es el mismo nº de filas que hay en la variable Otros de la columna "country_id". Es posible que Otros no tenga "region_code", por lo tanto, podemos eliminar la variable Otros.

# Filtrar el DataFrame para obtener solo las filas donde 'country_id' no sea 'ES'
df_otros = df_socialdemographic[df_socialdemographic['country_id'] == 'Otros']

# Contar las ocurrencias de 'region_code' en el subconjunto filtrado
region_code_counts = df_otros['region_code'].value_counts()

# Mostrar el resultado
print(region_code_counts)


# Imputar nulos en la columna 'salary' agrupando por 'age'
df_socialdemographic['salary'] = df_socialdemographic.groupby('age')['salary'].transform(lambda series: series.bfill().ffill())

df_socialdemographic.reset_index(drop=True,inplace=True)

df_socialdemographic.isnull().sum()

# Ordenar dataframe por fecha de más atigua a más nueva

# Ordenar el DataFrame por la columna 'pk_partition' de más antiguo a más nuevo
df_socialdemographic = df_socialdemographic.sort_values(by='pk_partition', ascending=True)
df_products = df_products.sort_values(by='pk_partition', ascending=True)
df_comercial_activity = df_comercial_activity.sort_values(by='pk_partition', ascending=True)

df_socialdemographic["pk_partition"].unique()

df_products["pk_partition"].unique()

df_comercial_activity["pk_partition"].unique()

# Merge

Antes de seguir haciendo limpieza hay que hacer un merge de los 3 deaframes

# Fusionar df_comercial_activity y df_products por 'pk_cid' 
df_merged = pd.merge(df_comercial_activity, df_products, on='pk_cid', how='inner', suffixes=('_comercial', '_product'))

df_merged.info()

#Vamos a borrar la columna "pk_partition_product" porque está duplicada y a parte es type object
df_merged.drop(columns=['pk_partition_product'], inplace=True)

df_merged = df_merged.rename(columns={'pk_partition_comercial': 'pk_partition'})
#Cambiamos el nombre de la columna para dejar el nombre estandar de antes

df_merged.info()

# Fusionar el resultado con df_socialdemographic por 'pk_cid'
df_final = pd.merge(df_merged, df_socialdemographic, on='pk_cid', how='inner', suffixes=('_merged', '_social'))

#Vamos a borrar la columna "pk_partition_product" porque está duplicada y a parte es type object
df_final.drop(columns=['pk_partition_merged'], inplace=True)

df_final = df_final.rename(columns={'pk_partition_social': 'pk_partition'})
#Cambiamos el nombre de la columna para dejar el nombre estandar de antes

df_final.info()

df_final.isnull().sum()

# Exportamos el df_final a un archivo .csv para hacer la segmentación

# Guardar resultados
df_final.to_csv('df_preprocessing.csv', index=False)

## Creamos un nuevo df por tipo de producto para hacer el encoding y posterior modelaje
