==> NOTEBOOKS/2.-encoding.ipynb <==
# Librerias

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder,MinMaxScaler, OrdinalEncoder
import random
pd.set_option('display.max_columns',None)

import warnings
warnings.filterwarnings("ignore")
from sklearn.preprocessing import OrdinalEncoder

import warnings
warnings.filterwarnings("ignore")

# fechas
from datetime import datetime

# gráficos
import seaborn as sns
import matplotlib.pyplot as plt

# preprocessing
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, Normalizer, MinMaxScaler
from sklearn.preprocessing import PolynomialFeatures

# algoritmos de clasificación
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression, RidgeClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from xgboost import XGBClassifier
from catboost import CatBoostClassifier

# algoritmos de regresión
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet, SGDRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from xgboost import XGBRegressor
from catboost import CatBoostRegressor

# métricas
import sklearn.metrics as metrics

# hiperparametrización
from sklearn.model_selection import GridSearchCV

# TensorFlow y tf.keras
import tensorflow as tf
from tensorflow import keras

import xgboost as xgb

from numpy import random
random.seed(42)

from sklearn.model_selection import train_test_split

Cargamos los datos

file_path = "C:/Users/Herre/Desktop/TFM/Nuclio_EasyMoney/0_Data/df_preprocessing.csv"
df_encoding = pd.read_csv(file_path)

df_encoding

df_encoding.isnull().sum()

df_encoding.info()

#Nos hemos dado cuenta de que las columnas "entry_date" y "pk_partition" se han vuelto a convertir en tipo Object
df_encoding['entry_date'] = pd.to_datetime(df_encoding['entry_date'], errors='coerce', format='%Y-%m-%d')
df_encoding['pk_partition'] = pd.to_datetime(df_encoding['pk_partition'], errors='coerce', format='%Y-%m-%d')

Necesitamos que todas las columnas sean numéricas, así que vamos a convertir las columnas de fechas a numéricas pero sin perder información.

# Crear columnas separadas para año, mes y día
df_encoding['entry_date_year'] = df_encoding['entry_date'].dt.year
df_encoding['entry_date_month'] = df_encoding['entry_date'].dt.month
df_encoding['entry_date_day'] = df_encoding['entry_date'].dt.day
df_encoding['entry_date_weekday'] = df_encoding['entry_date'].dt.weekday  # 0=Monday, 6=Sunday

df_encoding['pk_partition_year'] = df_encoding['pk_partition'].dt.year
df_encoding['pk_partition_month'] = df_encoding['pk_partition'].dt.month
df_encoding['pk_partition_day'] = df_encoding['pk_partition'].dt.day
df_encoding['pk_partition_weekday'] = df_encoding['pk_partition'].dt.weekday  # 0=Monday, 6=Sunday

#Como ya hemos creado una columna para años, mes, dia y día de la semana. Podemos elimianr la columna original
df_encoding.drop(columns=['entry_date'], inplace=True)
df_encoding.drop(columns=['pk_partition'], inplace=True)

# One Hot Encoding

df_encoding.info()

Columnas_OHE = ["entry_channel", "country_id", "gender", "deceased"]

# Convertir las variables categóricas en variables dummy
df_encoding = pd.get_dummies(data=df_encoding, columns=Columnas_OHE, dtype=int)

df_encoding

# Ordinal Encoding

Suponemos que los números junto a las variables de las columnas nos dicen la importancia de las mismas

# Crear un diccionario para el mapeo de valores
ordinal_mapping = {
    'Unknown': 0,
    '03 - UNIVERSITARIO': 1,
    '02 - PARTICULARES': 2,
    '01 - TOP': 3
}

# Aplicar el mapeo a la columna 'segment'
df_encoding['segment'] = df_encoding['segment'].map(ordinal_mapping)

## Borrar columnas duales

# Eliminar las columnas especificadas
df_encoding.drop(columns=["gender_H", "deceased_N"], inplace=True)

df_encoding

# Exportamos el archivo para pasar a la segmentación

# Guardar resultados
df_encoding.to_csv('df_encoding.csv', index=False)
