==> NOTEBOOKS/4.-agrupacion.ipynb <==
import pandas as pd

Cargamos los datos con el dataframe que salió al hacer el Encoding, necesitamos un dataframe sin las nuevas Features que creamos en la Segmentación porque nos afecta de manera negativa a la hora de hacer los modelos de Clasificación.

# Carga datos desde aquí

ruta_archivo = "C:/Users/Herre/Desktop/TFM/Nuclio_EasyMoney/0_Data/df_encoding.csv"

# Leer el archivo
df_agrupacion = pd.read_csv(ruta_archivo)

Agrupamos los productos en 3 grupos diferentes para poder hacer el modelo de Clasificación

# Categoría de Ahorro e Inversión
df_agrupacion['category_investment'] = (
    df_agrupacion['short_term_deposit'] +
    df_agrupacion['long_term_deposit'] +
    df_agrupacion['funds'] +
    df_agrupacion['securities'] +
    df_agrupacion['pension_plan']
)

# Categoría de Financiación
df_agrupacion['category_financing'] = (
    df_agrupacion['loans'] +
    df_agrupacion['mortgage'] +
    df_agrupacion['credit_card']
)

# Categoría de Cuentas
df_agrupacion['category_accounts'] = (
    df_agrupacion['em_account_pp'] +
    df_agrupacion['emc_account'] +
    df_agrupacion['em_account_p'] +
    df_agrupacion['em_acount'] +
    df_agrupacion['payroll_account'] +
    df_agrupacion['payroll'] +
    df_agrupacion['debit_card']
)

df_agrupacion

# Convertir a binario: 1 si el cliente tiene al menos un producto en la categoría, 0 si no tiene ninguno
df_agrupacion['category_investment'] = (df_agrupacion['category_investment'] > 0).astype(int)
df_agrupacion['category_financing'] = (df_agrupacion['category_financing'] > 0).astype(int)
df_agrupacion['category_accounts'] = (df_agrupacion['category_accounts'] > 0).astype(int)

# Eliminar columnas originales si ya no son necesarias
df_agrupacion.drop(['short_term_deposit', 'long_term_deposit', 'funds', 'securities',
                  'pension_plan', 'loans', 'mortgage', 'credit_card', 'em_account_pp',
                  'emc_account', 'em_account_p', 'em_acount', 'payroll_account', 
                  'payroll', 'debit_card'], axis=1, inplace=True)

df_agrupacion.info()

Revisamos las distribuciones de los 3 grupos de productos.

# Revisar la distribución de la variable objetivo
df_agrupacion['category_investment'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

# Revisar la distribución de la variable objetivo
df_agrupacion['category_financing'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

# Revisar la distribución de la variable objetivo
df_agrupacion['category_accounts'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

df_agrupacion.info()

Nos damos cuenta de que las distribuciones de los grupos de productos están muy descompensadas, es por eso que los datos se deben "balancear"

df_balanceo = df_agrupacion.copy()

# Guardar resultados
df_balanceo.to_csv('df_balanceo.csv', index=False)
