==> NOTEBOOKS/5.-balancear.ipynb <==
# Balancear Datos

import pandas as pd

ruta_archivo = "C:/Users/Herre/Desktop/Entrega_TFM/TFM_EasyMoney/df_balanceo.csv"

# Leer el archivo
df_balanceo = pd.read_csv(ruta_archivo)

df_balanceo

Anteriormente nos hemos dado cuenta de que los datos en los 3 grupos de productos están muy descompensados.

La gran mayoría de los usuarios No tienen los productos que ofrece el banco.

Si no balanceamos los datos el porcentaje de acierto de los modelos se dispara debido al overfitting. Salen porcentajes de más del 98%.

# Investment

# Revisar la distribución de la variable objetivo
df_balanceo['category_investment'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

# Contar las ocurrencias de cada clase
conteo_clases = df_balanceo['category_investment'].value_counts()

# Mostrar el número de '1'
num_1 = conteo_clases[1]
print(f"Número de filas con '1': {num_1}")

# Número de muestras de la clase "1" en df_balanceo
num_1 = df_balanceo[df_balanceo['category_investment'] == 1].shape[0]

# Seleccionar de manera aleatoria el mismo número de muestras de la clase "0"
df_0_random = df_balanceo[df_balanceo['category_investment'] == 0].sample(n=num_1, random_state=42)

# Crear el dataframe filtrado solo con los valores 1
df_investment = df_balanceo[df_balanceo['category_investment'] == 1]

# Unir ambos dataframes para tener uno balanceado
df_investment = pd.concat([df_investment, df_0_random])

# Mezclar las filas de manera aleatoria para evitar orden
df_investment = df_investment.sample(frac=1, random_state=42).reset_index(drop=True)

# df_investment ahora contiene el mismo número de "0" y "1" en la variable "category_investment"

# Revisar la distribución de la variable objetivo
df_investment['category_investment'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

df_investment.reset_index(drop=True,inplace=True)

### Retirar columnas

Esto afectaba negativamente

# Eliminar las columnas category_financing y category_accounts
df_investment = df_investment.drop(columns=['category_financing', 'category_accounts'])

# Verificamos que las columnas hayan sido eliminadas
print(df_investment.columns)

modelo_investment = df_investment.copy() 

# Guardar resultados
modelo_investment.to_csv('modelo_investment.csv', index=False)

# Financing

# Revisar la distribución de la variable objetivo
df_balanceo['category_financing'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

# Contar las ocurrencias de cada clase
conteo_clases = df_balanceo['category_financing'].value_counts()

# Mostrar el número de '1'
num_1 = conteo_clases[1]
print(f"Número de filas con '1': {num_1}")

# Número de muestras de la clase "1" en df_balanceo
num_1 = df_balanceo[df_balanceo['category_financing'] == 1].shape[0]

# Seleccionar de manera aleatoria el mismo número de muestras de la clase "0"
df_0_random = df_balanceo[df_balanceo['category_financing'] == 0].sample(n=num_1, random_state=42)

# Crear el dataframe filtrado solo con los valores 1
df_financing = df_balanceo[df_balanceo['category_financing'] == 1]

# Unir ambos dataframes para tener uno balanceado
df_financing = pd.concat([df_financing, df_0_random])

# Mezclar las filas de manera aleatoria para evitar orden
df_financing = df_financing.sample(frac=1, random_state=42).reset_index(drop=True)

# df_investment ahora contiene el mismo número de "0" y "1" en la variable "category_investment"

# Revisar la distribución de la variable objetivo
df_financing['category_financing'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

df_financing.info()

df_financing.reset_index(drop=True,inplace=True)

modelo_financing = df_financing.copy() 

# Guardar resultados
modelo_financing.to_csv('modelo_financing.csv', index=False)

# Account

df_balanceo.info()

# Revisar la distribución de la variable objetivo
df_balanceo['category_accounts'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

# Contar las ocurrencias de cada clase
conteo_clases = df_balanceo['category_accounts'].value_counts()

# Mostrar el número de '0'
num_0 = conteo_clases[0]
print(f"Número de filas con '0': {num_0}")

# Número de muestras de la clase "0" en df_balanceo
num_0 = df_balanceo[df_balanceo['category_accounts'] == 0].shape[0]

# Seleccionar de manera aleatoria el mismo número de muestras de la clase "1"
# Añadir replace=True para permitir la selección con reemplazo si es necesario
df_1_random = df_balanceo[df_balanceo['category_accounts'] == 1].sample(n=num_0, replace=True, random_state=42)

# Crear el dataframe filtrado solo con los valores 0
df_account = df_balanceo[df_balanceo['category_accounts'] == 0]

# Unir ambos dataframes para tener un conjunto balanceado
df_account = pd.concat([df_account, df_1_random])

# Mezclar las filas de manera aleatoria para evitar orden
df_account = df_account.sample(frac=1, random_state=42).reset_index(drop=True)

# Ahora df_account contiene el mismo número de "0" y "1" en la variable 'category_accounts'

# Revisar la distribución de la variable objetivo
df_account['category_accounts'].value_counts(normalize=True) # Reemplaza TARGET con tu columna real

df_account.info()

df_account.reset_index(drop=True,inplace=True)

modelo_account = df_account.copy() 

modelo_account.info()

# Guardar resultados
modelo_account.to_csv('modelo_account.csv', index=False)
