==> NOTEBOOKS/6.2.-Modelo_account.ipynb <==
# Librerías

# cálculos
import numpy as np
import pandas as pd
import math

# fechas
from datetime import datetime

# gráficos
import seaborn as sns
import matplotlib.pyplot as plt

# preprocessing
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split

# modelos
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression, RidgeClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
from sklearn.metrics import accuracy_score, roc_auc_score

# metricas
import sklearn.metrics as metrics

# hiperparametrizado
from sklearn.model_selection import GridSearchCV

# quitar alertas
import warnings
warnings.filterwarnings("ignore")



 # Ruta al archivo 
ruta_archivo = "C:/Users/Herre/Desktop/Entrega_TFM/TFM_EasyMoney/modelo_account.csv"
# Leer el archivo 
df_modelo = pd.read_csv(ruta_archivo)

# Separar las características (X) y la variable objetivo (y)
X = df_modelo.drop(columns=['category_accounts'])  # Todas las columnas excepto 'category_accounts'
y = df_modelo['category_accounts']  # La columna 'category_accounts' es la variable objetivo

# Dividir los datos (80% entrenamiento, 20% prueba)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Toreno de modelos

f1s = pd.DataFrame(columns=['modelo', 'f1_score'])
f1s

# añadimos en una lista los modelos que queremos poner a competir

modelos = []
modelos.append(('LogisticRegression', LogisticRegression()))
modelos.append(('RidgeClassifier', RidgeClassifier()))
modelos.append(('DecisionTreeClassifier', DecisionTreeClassifier()))
modelos.append(('RandomForestClassifier', RandomForestClassifier()))
modelos.append(('GradientBoostingClassifier', GradientBoostingClassifier()))
modelos.append(('XGBClassifier', XGBClassifier()))
modelos.append(('CatBoostClassifier', CatBoostClassifier(verbose=False)))

for nombre, modelo in modelos:
  # entrenamiento
  modelo.fit(X_train, y_train)

  # predicción del test
  y_pred = modelo.predict(X_test)

  # evalua la metrica
  f1_score = metrics.f1_score(y_test, y_pred, average='macro')
  #el average='macro' se usa cuando estamos haciendo modelos multiclases. Sino lo pones el modelo peta.

  # añade la métrica al dataframe de métricas
  metrica = pd.DataFrame({'modelo':[nombre], 'f1_score':[f1_score]})

  print(nombre, np.round(f1_score,5))

  f1s = pd.concat([f1s,metrica], ignore_index=True)

f1s.sort_values('f1_score', ascending= False)

# Modelo

# Ajuste de hiperparámetros
xgb_model = XGBClassifier(n_estimators=100, max_depth=3, learning_rate=0.1, use_label_encoder=False, eval_metric='logloss')

# Entrenar el modelo con los datos de entrenamiento
xgb_model.fit(X_train, y_train)

# Hacer predicciones con el conjunto de prueba
y_pred = xgb_model.predict(X_test)
y_pred_proba = xgb_model.predict_proba(X_test)[:, 1]  # Probabilidad de la clase 1 (compra)

from sklearn.metrics import accuracy_score, confusion_matrix, classification_report

# Calcular la precisión
accuracy = accuracy_score(y_test, y_pred)
print(f"Precisión: {accuracy * 100:.2f}%")

# Mostrar la matriz de confusión
conf_matrix = confusion_matrix(y_test, y_pred)
print("Matriz de confusión:")
print(conf_matrix)

# Mostrar el reporte de clasificación
class_report = classification_report(y_test, y_pred)
print("Reporte de clasificación:")
print(class_report)

y_pred

y_pred_proba

# Crear un nuevo DataFrame con 'pk_cid', las predicciones y las probabilidades

df_resultado = X_test[['pk_cid']].copy()  # Copiar la columna 'pk_cid' de X_test
df_resultado['prediccion'] = y_pred  # Agregar las predicciones
df_resultado['probabilidad_compra'] = y_pred_proba  # Agregar la probabilidad de compra

# Mostrar el nuevo DataFrame
print(df_resultado.head())  # Muestra las primeras filas del DataFrame

df_resultado.info()

En este dataframe tenemos muchos clientes, por lo tanto vamos a filtrar por aquellos que tienen un probabilidad de compra de casi el 100%

# Filtrar filas donde la probabilidad de compra sea mayor al 96.5%
df_filtrado = df_resultado[df_resultado['probabilidad_compra'] > 0.965]

df_filtrado

Para el siguiente paso del proyecto nos piden 10 mil clientes por lo tanto no podemos coger a todos los clientes, cogemos el número de clientes que nos faltan para llegar a 10 mil.

# Ordenar el dataframe por la columna 'probabilidad_compra' en orden descendente
df_ordenado = df_filtrado.sort_values(by='probabilidad_compra', ascending=False)

# Seleccionar las primeras 4000 filas
df_top_4000 = df_ordenado.head(4076)

df_top_4000

Exportamos la prediccion porque la necesitaremos más adelante para conocer los ingresos que tendrá la empresa

# Guardar el DataFrame en un archivo CSV para revisarlo luego
df_top_4000.to_csv('accounts_resultado_predicciones.csv', index=False)
